requirements.txt
python-dotenv
asyncpg
aiogram
aiohttp
aioredis
Authlib
requests